+++ app/[locale]/template/[templateId]/access/page.tsx
@@ -15,25 +15,23 @@ import {
 // Graphql mutations
 import {
   TemplateCollaboratorsDocument,
-  useTemplateCollaboratorsQuery,
   useAddTemplateCollaboratorMutation,
-  useRemoveTemplateCollaboratorMutation
+  useRemoveTemplateCollaboratorMutation,
+  useTemplateCollaboratorsQuery
 } from '@/generated/graphql';
 
 // Components
 import PageHeader from "@/components/PageHeader";
 import FormInput from '@/components/Form/FormInput';
 import ConfirmModal from '@/components/Modal/ConfirmModal';
-import {
-  LayoutContainer,
-  ContentContainer,
-} from '@/components/Container';
+import { ContentContainer, LayoutContainer, } from '@/components/Container';
 
 //Utils and other
 import logECS from '@/utils/clientLogger';
 import { isValidEmail } from '@/utils/validation';
 import { scrollToTop } from '@/utils/general';
 import { useToast } from '@/context/ToastContext';
+
 import styles from './TemplateAccessPage.module.scss';
 
 const GET_COLLABORATORS = TemplateCollaboratorsDocument;
@@ -136,8 +134,8 @@ const TemplateAccessPage: React.FC = () => {
       });
 
       const emailData = response?.data?.addTemplateCollaborator;
-      if (emailData?.errors?.length) {
-        return setErrorMessages(emailData.errors);
+      if (emailData?.errors && Object.values(emailData?.errors).filter((err) => err && err !== 'TemplateCollaboratorErrors').length > 0) {
+        setErrorMessages([emailData?.errors.general || AccessPage('messages.errors.errorAddingCollaborator')]);
       }
 
       clearErrors();
@@ -265,7 +263,7 @@ const TemplateAccessPage: React.FC = () => {
           <div className="template-editor-container" ref={errorRef}>
             <div className="main-content">
               {errorMessages && errorMessages.length > 0 &&
-                <div className="error" role="alert" aria-live="assertive">
+                <div className="messages error" role="alert" aria-live="assertive" ref={errorRef}>
                   {errorMessages.map((error, index) => (
                     <p key={index}>{error}</p>
                   ))}