+++ app/[locale]/template/[templateId]/q/[q_slug]/page.tsx
@@ -1,9 +1,9 @@
 'use client'
 
-import { useEffect, useRef, useState } from 'react';
-import { ApolloError } from '@apollo/client';
-import { useParams, useRouter, useSearchParams } from 'next/navigation';
-import { useTranslations } from 'next-intl';
+import {useEffect, useRef, useState} from 'react';
+import {ApolloError} from '@apollo/client';
+import {useParams, useRouter, useSearchParams} from 'next/navigation';
+import {useTranslations} from 'next-intl';
 import {
   Breadcrumb,
   Breadcrumbs,
@@ -24,20 +24,21 @@ import {
 // GraphQL queries and mutations
 import {
   useQuestionQuery,
-  useUpdateQuestionMutation,
-  useQuestionTypesQuery
+  useQuestionTypesQuery,
+  useUpdateQuestionMutation
 } from '@/generated/graphql';
 
 // Components
 import PageHeader from "@/components/PageHeader";
-import QuestionOptionsComponent from '@/components/Form/QuestionOptionsComponent';
+import QuestionOptionsComponent
+  from '@/components/Form/QuestionOptionsComponent';
 import FormInput from '@/components/Form/FormInput';
 import FormTextArea from '@/components/Form/FormTextArea';
 import ErrorMessages from '@/components/ErrorMessages';
 
 //Other
-import { useToast } from '@/context/ToastContext';
-import { Question, QuestionOptions } from '@/app/types';
+import {useToast} from '@/context/ToastContext';
+import {Question, QuestionOptions} from '@/app/types';
 import styles from './questionEdit.module.scss';