+++ components/Form/FormTextArea/index.tsx
@@ -6,7 +6,6 @@ import {
   TextArea,
   TextField,
 } from "react-aria-components";
-import styles from './formInput.module.scss';
 
 interface InputProps {
   name: string;