+++ jest.config.ts
@@ -29,10 +29,10 @@ const config: Config = {
     ],
     coverageThreshold: {
         global: {
-            branches: 60,
-            functions: 60,
-            lines: 60,
-            statements: 60,
+            branches: 50,
+            functions: 50,
+            lines: 50,
+            statements: 50,
         }
     },
     coverageDirectory: "coverage",