+++ lib/graphql/apollo-wrapper.tsx
@@ -1,20 +1,16 @@
 "use client"
 
-import { createHttpLink, ApolloLink } from "@apollo/client";
-import { errorLink, authLink, retryLink } from "@/lib/graphql/graphqlHelper";
+import {ApolloLink, createHttpLink} from "@apollo/client";
+import {authLink, errorLink, retryLink} from "@/lib/graphql/graphqlHelper";
 
 //From https://github.com/apollographql/apollo-client-nextjs
-
 import {
-    ApolloNextAppProvider,
+  ApolloClient,
+  ApolloNextAppProvider,
+  InMemoryCache,
 } from "@apollo/experimental-nextjs-app-support";
 
-import {
-    ApolloClient,
-    InMemoryCache
-} from "@apollo/experimental-nextjs-app-support";
-
-import { setApolloClient } from './apolloClient';
+import {setApolloClient} from './apolloClient';
 
 function makeClient() {
     const httpLink = createHttpLink({